Split authRouter into two seperate functions under auth to be autoloaded
